// web/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const DayOfWeek = {
  MONDAY: 'MONDAY',
  TUESDAY: 'TUESDAY',
  WEDNESDAY: 'WEDNESDAY',
  THURSDAY: 'THURSDAY',
  FRIDAY: 'FRIDAY',
  SATURDAY: 'SATURDAY',
  SUNDAY: 'SUNDAY'
} as const

export type DayOfWeek = (typeof DayOfWeek)[keyof typeof DayOfWeek]


export const AttendanceStatus = {
  PENDING: 'PENDING',
  PRESENT: 'PRESENT',
  LATE: 'LATE',
  ABSENT: 'ABSENT',
  PERMISSION: 'PERMISSION',
  SICK: 'SICK'
} as const

export type AttendanceStatus = (typeof AttendanceStatus)[keyof typeof AttendanceStatus]


export const CheckInStatus = {
  ON_TIME: 'ON_TIME',
  LATE: 'LATE'
} as const

export type CheckInStatus = (typeof CheckInStatus)[keyof typeof CheckInStatus]


export const CheckOutStatus = {
  NORMAL: 'NORMAL',
  EARLY: 'EARLY'
} as const

export type CheckOutStatus = (typeof CheckOutStatus)[keyof typeof CheckOutStatus]